supporting-blog-content/building-a-recipe-search-with-elasticsearch/requirements.txt
elasticsearch
aiohttp
pyyaml
pandas
load_dotenv
